add header guard
The current implementation does not contain guard and it works well,
but it may lead to problems in the future.

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>

add a subtraction operation

Add member-function which performs a subtraction in terms of
an addition.

Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>

improve calculation accuracy

Allow using float point arguments to avoid truncation.

Test: Add (4.9, 4.9) should return 10.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>

fix truncation error

To convert float to integer the truncation is performed, but the
rounding is expected.

Test: Add (4.9, 4.9) should return 10.
Signed-off-by: Sergii Piatakov <sergii.piatakov@globallogic.com>